style : 컴포넌트 경로 수정 - component 폴더 아래에 posting 폴더로 이동 - style 폴더 아래에 theme/theme.js 파일에 색상 작성
ref : #5
