+++ src/components/index.js
@@ -1,14 +1,14 @@
-import InputField from "./InputField";
-import Tag from "./Tag";
-import Toggle from "./Toggle";
-import Button from "./Button";
-import WorkTimePicker from "./WorkTimepicker";
-import WeekdayPicker from "./WeekdayPicker";
-import PayPicker from "./PayPicker";
-import AddressInput from "./AddressInput";
-import PhotoUpload from "./PhotoUpload";
-import DescriptionInput from "./DescriptionInput";
-import PhoneInput from "./PhoneInput";
+import InputField from "./posting/InputField";
+import Tag from "./posting/Tag";
+import Toggle from "./posting/Toggle";
+import Button from "./posting/Button";
+import WorkTimePicker from "./posting/WorkTimepicker";
+import WeekdayPicker from "./posting/WeekdayPicker";
+import PayPicker from "./posting/PayPicker";
+import AddressInput from "./posting/AddressInput";
+import PhotoUpload from "./posting/PhotoUpload";
+import DescriptionInput from "./posting/DescriptionInput";
+import PhoneInput from "./posting/PhoneInput";
 
 export {InputField, Tag,
     Toggle, Button, WeekdayPicker, WorkTimePicker, PayPicker, AddressInput,
