--- src/constants/index.js
@@ -1,8 +1,0 @@
-const colors={
-    carrot: "#ff8a3d",
-    green: "#339f46",
-    검정:"#000000",
-    연회색:"cccccc",
-    흰색:"ffffff"
-
-}
