+++ src/styles/theme/theme.js
@@ -1,0 +1,8 @@
+const colors={
+    carrot: "#ff8a3d",
+    green: "#339f46",
+    검정:"#000000",
+    연회색:"cccccc",
+    흰색:"ffffff"
+
+}
